split tests out into two separate files
